fix minors class modifiers
