+++ lib/base_component.dart
@@ -1,6 +1,6 @@
 library base_project;
 
-// export 'src/base_component/light_observer.dart' show LightObserver, MultiObserWidget;
 export 'src/base_component/base_controller.dart' show DefaultController;
 export 'src/state_management/extension.dart' show Global;
-export 'src/base_component/base_observer.dart' show Observer, ObserWidget, ObserListWidget, ObserverCombined;
+export 'src/base_component/base_observer.dart'
+    show Observer, ObserWidget, ObserListWidget, ObserverCombined;
