+++ lib/base_widget.dart
@@ -1,6 +1,6 @@
 library base_project;
 
-export 'src/widgets/custom_snackbar.dart' show AppSnackBar, SnackBarStyle;
+export 'src/interfaces/widget_interfaces.dart' show SnackBarStyle;
 export 'src/widgets/main_widget.dart';
-export 'src/widgets/custom_loading.dart';
-export 'src/nav_dialog/custom_dialog.dart' show AppDialog;
+export 'src/nav_dialog/public_dialog.dart' show AppDialog;
+export 'src/widgets/public_widget.dart';
