+++ lib/src/interfaces/observer_interfaces.dart
@@ -1,0 +1,11 @@
+abstract interface class ObserverAbs<T> {
+  void update();
+
+  Stream<T> get stream;
+
+  T get value;
+
+  set value(T valueSet);
+
+  dispose();
+}
